rhoPimpleFoam: Improved support for compressible liquids
Updates solver for changes to OpenFOAM-dev branch
https://github.com/OpenFOAM/OpenFOAM-dev/commit/79ff91350ef6092b2dd864029a6023bec2b66b50

"
rhoPimpleFoam: Improved support for compressible liquids
See tutorials/compressible/rhoPimpleFoam/RAS/squareBendLiq for exapmle

pimpleControl: Added SIMPLErho option for running in SIMPLE mode

with large time-step/Courant number and relaxation.  With this option the
density is updated from thermodynamics rather than continuity after the pressure
equation which is better behaved if pressure is relaxed and/or solved to a
loose relative tolerance.  The need for this option is demonstrated in the
tutorials/compressible/rhoPimpleFoam/RAS/angledDuct tutorial which is unstable
without the option.


"